Add ImGui and ImGuizmo, add first implementation of a Gui currently only rendering sample window. Add gui settings and gui shaders. Event system now allows an early subscription via subscribe_early. Mesh has a way to update its data. Model accepts more complete primitve description to create default mesh and provide a reset function. Scene adds rendering of Gui as the last element in the draw order. Renderer now accepts a dimension and scale inside dimensions. Resource types are updated to reflect the other types correctly. Buffer also provides a resize method now. Scissor and buffer offset functionality added to encoders. Add build_shader utility function. Lots of settings added.

+++ core/camera/rorcamera.cpp
@@ -159,7 +159,7 @@ void OrbitCamera::update_projection()
 	if (this->m_type == CameraType::perspective)
 	{
 		// assumes square pixels, if ever there is evidence of non-square pixels, provided by windowing system i.e. glfw use it here instead
-		this->m_projection = ror::make_perspective(ror::to_radians(this->m_y_fov), static_cast<float>(this->m_width) / static_cast<float>(this->m_height), this->m_z_near, this->m_z_far);
+		this->m_projection = ror::make_perspective(ror::to_radians(this->m_y_fov), this->m_width / this->m_height, this->m_z_near, this->m_z_far);
 
 		// Make infinite projections matrix
 		float a_z_near = 0.1f;
@@ -169,7 +169,7 @@ void OrbitCamera::update_projection()
 		this->m_projection.m_values[14] = (epsilon - 2.0f) * a_z_near;
 	}
 	else
-		this->m_projection = ror::make_ortho(0.0f, static_cast<float>(this->m_width), 0.0f, static_cast<float>(this->m_height), this->m_z_near, this->m_z_far);
+		this->m_projection = ror::make_ortho(0.0f, this->m_width, 0.0f, this->m_height, this->m_z_near, this->m_z_far);
 }
 
 void OrbitCamera::setup()
